+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Downloadable releases](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- core search place name in Fulltext
 - plugin import K.Import.geojsonToPlace()
 
 ## v1.6.6